+++ ChangeLog.md
@@ -7,10 +7,10 @@ More detailed information about incremental changes can be found in the
 
 ## 3.14
 
-### 3.14.10 (2022-xx-yy)
+### 3.14.10 (2022-10-11)
 
-- Added option grad_f_constant to specify that objective function is linear.
-  If set, the gradient of the objective will be requested by NLP only once. [#597]
+- Added option `grad_f_constant` to specify that objective function is linear.
+  If set, the gradient of the objective will be requested by Ipopt only once. [#597]
 - Added `OrigIpoptNLP::orig_d_L()` and `OrigIpoptNLP::orig_d_U()` to get
   original constraint sides (before relaxation due to bound_relax_factor > 0).
 - `TNLP::get_curr_violations()` now returns the constraint violation and